Moved options-info to configs folder
